+++ build.xml
@@ -246,8 +246,8 @@ Copyright (c) 2005, 2017, Oracle and/or its affiliates. All rights reserved.
         <run-jflex dir="${gen.analysis.dir}/javascript" name="JavaScriptXref"/>
         <run-jflex dir="${gen.analysis.dir}/lua" name="LuaSymbolTokenizer"/>
         <run-jflex dir="${gen.analysis.dir}/lua" name="LuaXref"/>
-        <run-jflex dir="${gen.analysis.dir}/posh" name="PoshSymbolTokenizer"/>
-        <run-jflex dir="${gen.analysis.dir}/posh" name="PoshXref"/>
+        <run-jflex dir="${gen.analysis.dir}/powershell" name="PowershellSymbolTokenizer"/>
+        <run-jflex dir="${gen.analysis.dir}/powershell" name="PowershellXref"/>
         <run-jflex dir="${gen.analysis.dir}/python" name="PythonSymbolTokenizer"/>
         <run-jflex dir="${gen.analysis.dir}/python" name="PythonXref"/>
         <run-jflex dir="${gen.analysis.dir}/rust" name="RustSymbolTokenizer"/>
