+++ nbproject/project.properties
@@ -94,7 +94,7 @@ javadoc.encoding=${source.encoding}
 build.test.classes.dir=${build.dir}/test/classes
 jar.index=${jnlp.enabled}
 file.reference.jrcs.jar=lib/jrcs.jar
-javac.compilerargs=-Xlint:-fallthrough
+javac.compilerargs=-Xlint:unchecked 
 source.encoding=UTF-8
 jnlp.signing.keystore=
 meta.inf.dir=${src.dir}/META-INF
