+++ src/org/opensolaris/opengrok/analysis/AnalyzerGuru.java
@@ -82,7 +82,7 @@ import org.opensolaris.opengrok.analysis.clojure.ClojureAnalyzerFactory;
 import org.opensolaris.opengrok.analysis.json.JsonAnalyzerFactory;
 import org.opensolaris.opengrok.analysis.kotlin.KotlinAnalyzerFactory;
 import org.opensolaris.opengrok.analysis.sh.ShAnalyzerFactory;
-import org.opensolaris.opengrok.analysis.posh.PoshAnalyzerFactory;
+import org.opensolaris.opengrok.analysis.powershell.PowershellAnalyzerFactory;
 import org.opensolaris.opengrok.analysis.sql.PLSQLAnalyzerFactory;
 import org.opensolaris.opengrok.analysis.sql.SQLAnalyzerFactory;
 import org.opensolaris.opengrok.analysis.swift.SwiftAnalyzerFactory;
@@ -178,7 +178,7 @@ public class AnalyzerGuru {
             new CxxAnalyzerFactory(),
             new ErlangAnalyzerFactory(),
             new ShAnalyzerFactory(),
-            new PoshAnalyzerFactory(),
+            new PowershellAnalyzerFactory(),
             PlainAnalyzerFactory.DEFAULT_INSTANCE,
             new UuencodeAnalyzerFactory(),
             new GZIPAnalyzerFactory(),
