+++ src/org/opensolaris/opengrok/analysis/powershell/PowershellSymbolTokenizer.lex
@@ -21,7 +21,7 @@
  * Copyright (c) 2017, Oracle and/or its affiliates. All rights reserved.
  */
 
-package org.opensolaris.opengrok.analysis.posh;
+package org.opensolaris.opengrok.analysis.powershell;
 import java.io.IOException;
 import java.io.Reader;
 import org.opensolaris.opengrok.analysis.JFlexTokenizer;
